ath9k: rework Ubuntu's MSI support for vector restriction
Ubuntu have added patches for ath9k MSI support via a module param
and DMI quirk list, which may be enough for the platforms in that list.

However they have not addressed the problem that we solve with our
other work here:

We observed that the device seems to unset the 2 lower bits of
the MSI Message Data, causing do_IRQ to report:
  0.64 No irq handler for vector

To work around this, we restrict ourselves to vector numbers that
are aligned by 4 (i.e. with the lower bits already unset), that way
the discrepency is avoided.

We limit this change to AR9565, since we had multiple user reports
of AR9462 having connection difficulties when MSI support is enabled,
and reproduced ping test instability in the Taipei office on Bobby's
AR9462 laptop when MSI interrupts were in use.

https://phabricator.endlessm.com/T16988
https://phabricator.endlessm.com/T19774