rtl8723bu: always indicate disconnect during netdev_close
In eos3.5, CONFIG_ANDROID=n, so we were running this code, which
we've previously documented as being the essential thing to tell
cfg80211 that we've disconnected in the suspend/resume path.

In eos3.6, CONFIG_ANDROID=y so we were no longer doing this.
After suspend/resume, wifi reconnection is broken as a result.

"Fix" this by restoring the eos3.5 behaviour of running this codepath.

https://phabricator.endlessm.com/T20835
https://phabricator.endlessm.com/T27164